Fixes #22489 - move the notification base class
